Initial work
This update allows all possible actions without being combinatorial with the possible discards.  Instead, we just have 4 moves per legal take, and discards will get automatically calculated.  Not sure if this will work, but can get progressively more hardcoded intelligence on the automatic discard if it doesn't so it's worth trying.

Main issue is that there is variance in some moves that does not exist in others, unstabilizing the reward signal by quite a bit.  However, all of these moves are rare and almost always suboptimal anyway.  Because the reward signal can only lower, and not increase, I suspect this will just act as a natural non-sparse reward signal to avoid these states.