State vector inclusions and normalization changes
Added a single dimension to the state vector for each player, the sum of the player purchased cards.  Changed normalization for many dimensions, moving to normalize gold separately.  Cards are also normalized to 10 to try to align all with what they actually do - cards being equivalent to gems as far as purchase math goes.