Vector length standardization, lr warmup
Previously the vectors were a mix of length 5 and 6.  Now nearly all of them are length 6, meaning many windows are removed and can instead be directly broadcast.

I also believe that the model is learning too heavily during the high exploration phase, so I implemented learning rate warmup.  If for no other reason than this can be a problem regardless.