First performant version, hitting 32-length games
Key change is a constant, low epsilon.  Will improve on this in later runs.  The model needs to be making coherent moves, as any random moves made after a good one will completely lock the model up and be unable to get its gems to where it wants later, and it may suffer the discard penalty.